pyproject.toml
[tool.poetry]
name = "lambdatest-py"
version = "0.1.0"
description = ""
authors = ["Joel Majors <joelamajors@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
autopep8 = "^1.5.7"
flake8 = "^3.9.2"
pyleniumio = "^1.12.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
